Added Clear all proto functionality

+++ app.py
@@ -1,21 +1,27 @@
-from flask import Flask, render_template, request, redirect, jsonify
+from flask import Flask, session, render_template, request, redirect, jsonify
 from flask_mysqldb import MySQL
-import yaml, json
+import yaml
 import random
 import time
 
 app = Flask(__name__)
+app.secret_key = 'tatanamak'
+
 # Configure db
 db = yaml.load(open('db.yaml'))
-app.config['MYSQL_USER'] = 'root'
-app.config['MYSQL_PASSWORD'] = ''
-app.config['MYSQL_HOST'] = 'localhost'
-app.config['MYSQL_DB'] = 'ssdproject'
+app.config['MYSQL_USER'] = db['mysql_user']
+app.config['MYSQL_PASSWORD'] = db['mysql_password']
+app.config['MYSQL_HOST'] = db['mysql_host']
+app.config['MYSQL_DB'] = db['mysql_db']
 mysql = MySQL(app)
 
 
-@app.route('/', methods=['GET', 'POST'])
+
+@app.route('/')
 def root():
+
+    session.pop('uid', None)
+    session.pop('wid', None)
     return render_template('index.html')
 
 
@@ -33,7 +39,7 @@ def userdetails():
     # inserting into database
     conn = mysql.connection
     cur = conn.cursor()
-    userid = cur.execute(
+    cur.execute(
         "INSERT INTO `user` (`username`,`name`,`password`,`college`,`type`) VALUES (%s,%s,%s,%s,%s)",
         queryValues)
     cur.close()
@@ -44,6 +50,11 @@ def userdetails():
 # inserting data into register table while logging in
 @app.route('/login', methods=['POST'])
 def login():
+
+
+    session.pop('uid', None)
+    session.pop('wid', None)
+    
     username = request.form.get('Email')
     password = request.form.get('Password')
 
@@ -51,41 +62,69 @@ def login():
     conn = mysql.connection
     cur = conn.cursor()
     values = cur.execute(
-        "SELECT user.userid,user.type  FROM user where user.username = %s AND user.password = %s",
+        "SELECT user.userid,user.type,user.username  FROM user where user.username = %s AND user.password = %s",
         [username, password])
     if (values > 0):
         userDetails = cur.fetchall()
         userid = userDetails[0][0]
         type = userDetails[0][1]
+        uname = userDetails[0][2]
+
+        #storing username and userid and grp id and workshop and type in session variable
+        session['uid'] = userid
+        # session['type'] = type
+        # session['uname']= uname
+        session['wid']= "1" #hardcoding the workshop Id
+
 
         # storing the userid in the registration table
-        
         queryValues = [1, userid]
-        cur.execute(
+        print(userid)
+
+        # checking if the user is loggedin atleast once
+        check_user = cur.execute("SELECT reg.userid FROM registration reg WHERE reg.userid = %s",(userid,))
+
+        # If not inserting his data into registration table
+        if(check_user == 0):
+            cur.execute(
             "INSERT INTO `registration` (`workshopid`, `userid`) VALUES (%s , %s)",
             queryValues)
-        cur.close()
-        conn.commit()
+            cur.close()
+            conn.commit()
     
         if (type == 'admin'):
             return redirect('/admin')
         elif (type == 'participant'):
-            return redirect('/emphasize')
+            return redirect('/waiting')
     else:
         return render_template('404.html')
+
+# logout
+@app.route('/logout')
+def logout():
+
+    session.pop('uid', None)
+    session.pop('wid', None)
+    return redirect('/')
 
 
 # functonalities of all admin
 @app.route('/admin', methods=['GET', 'POST'])
 def admin():
+    if not session.get("uid"):
+        return redirect("/")
     return render_template('admin.html')
 
 @app.route('/admin/teams', methods=['GET', 'POST'])
 def teams():
+    if not session.get("uid"):
+        return redirect("/")
     return render_template('teams.html')
 
 @app.route('/admin/people', methods=['GET', 'POST'])
 def people():
+    if not session.get("uid"):
+        return redirect("/")
     return render_template('participants.html')
 
 
@@ -93,6 +132,9 @@ def people():
 @app.route('/admin/people/showparticipants', methods=['post'])
 def showparticipants():
 
+
+    if not session.get("uid"):
+        return redirect("/")
     # connecting to database
     conn = mysql.connection
     cur = conn.cursor()
@@ -147,9 +189,12 @@ def arrangeRandom(userIdstuple, count):
 @app.route('/admin/createTeams', methods=["POST"])
 def createTeams():
 
-    # getting count of memebers in each team from form
-    # count = request.form.get('count')
 
+    if not session.get("uid"):
+        return redirect("/")
+    # getting count of memebers in each team from form
+    rsp = request.json
+    count = int(rsp['count'])
     # connecting to database
     conn = mysql.connection
     cur = conn.cursor()
@@ -162,7 +207,7 @@ def createTeams():
     if values > 0:
         
         userIds = cur.fetchall()
-        createdTeams = arrangeRandom(userIds, 3)
+        createdTeams = arrangeRandom(userIds, count)
         for key in createdTeams.keys():
              for value in createdTeams[key]:
                  cur.execute("INSERT INTO `group` (`workshopid`,`groupid`, `userid`) VALUES (%s,%s, %s)",(1,key,value))
@@ -186,11 +231,15 @@ def createTeams():
 # functonalities of all participants
 
 startTime = int(time.time())
-phaseTime = 120
+phaseTime = 3600
 
 
 @app.route("/updatePhase",methods=['GET'])
 def updatephase():
+
+    if not session.get("uid"):
+        return redirect("/")
+
     phase = request.args.get('curphase')
     
     #will fetch this time from DB which contains end time
@@ -231,9 +280,71 @@ def updatephase():
             return "false"
 
 
+
+#========================================================================
+#========================Prototype Part===============================
+#========================================================================
+
+
+@app.route("/returnProto",methods=['GET'])
+def returnProto():
+
+    if not session.get("uid"):
+        return redirect("/")
+
+    cur = mysql.connection.cursor()
+    query="select * from protoTable;"
+    cur.execute(query)
+    results = cur.fetchall()
+    
+    payload = []
+    content = {}
+    for result in results:
+        content = {'imgid': result[0], 'image': result[3]}
+        payload.append(content)
+        content = {}
+    
+
+    #print(payload);
+    cur.close()
+    return jsonify(payload)
+
+@app.route("/addProto",methods=['GET'])
+def addProto():
+
+    if not session.get("uid"):
+        return redirect("/")
+
+    cur = mysql.connection.cursor()
+    app.logger.info(request.args.keys())
+    wid = request.args.get('workid')
+    gid = request.args.get('grpid')
+    img = request.args.get('image')
+    
+    wid=1
+    gid=1
+
+    #cur.execute("select * from protoTable where ")
+
+    query="insert into protoTable(workshopid,grpid,image) values(%s,%s,%s);"
+    
+    record = [wid,gid,img]
+    cur.execute(query, record)
+    mysql.connection.commit()
+    cur.close()
+    return 'Proto Added Successfully'
+
+
+#========================================================================
+#========================Sticky Notes Part===============================
+#========================================================================
+
 @app.route("/deleteSticky",methods=['GET'])
 def deleteSticky():
     
+    if not session.get("uid"):
+        return redirect("/")
+
     cur = mysql.connection.cursor()
     uid = request.args.get('noteid')
     query="delete from wall where notesid="+uid+";"
@@ -244,6 +355,10 @@ def deleteSticky():
     
 @app.route("/returnSticky",methods=['GET'])
 def returnSticky():
+
+    if not session.get("uid"):
+        return redirect("/")
+
     cur = mysql.connection.cursor()
     query="select * from wall;"
     cur.execute(query)
@@ -262,9 +377,12 @@ def returnSticky():
     return jsonify(payload)
 
 
-
 @app.route("/addSticky",methods=['GET'])
 def addSticky():
+
+    if not session.get("uid"):
+        return redirect("/")
+
     cur = mysql.connection.cursor()
     wid = request.args.get('workid')
     gid = request.args.get('grpid')
@@ -280,9 +398,15 @@ def addSticky():
     cur.close()
     return 'Added Successfully'
     
+
+
     
 @app.route("/addMessage",methods=['GET'])
 def addMessage():
+
+    if not session.get("uid"):
+        return redirect("/")
+
     cur = mysql.connection.cursor()
     msg = request.args.get('msg')
     userid = request.args.get('userid')
@@ -298,6 +422,10 @@ def addMessage():
 
 @app.route("/getmsg")
 def getmsg():
+
+    if not session.get("uid"):
+        return redirect("/")
+
     cur = mysql.connection.cursor()
     query1="select * from chat order by messageid;"
     cur.execute(query1)
@@ -316,39 +444,181 @@ def getmsg():
     return jsonify(payload)
 
 
-
-@app.route("/verifylogin",methods=['POST'])
-def verifylogin():
-    return 'verifying login credentials'
-
 @app.route("/emphasize")
 def participants():
+
+    if not session.get("uid"):
+        return redirect("/")
+
     global startTime
     startTime = int(time.time())
-    return render_template('emphasize.html')
+
+
+
+
+    #*********** Run Sql Query To fetch wid, Gid Uid From DAtabase
+
+    wid = 1
+    gid = 1
+    uid = 2
+
+    data={"workshopid":wid, "groupid": gid, "userid" : uid}
+
+    return render_template('emphasize.html',user=data)
+
+@app.route("/waiting")
+def waiting():
+
+    if not session.get("uid"):
+        return redirect("/")
+
+    #useing session variable.........
+    if not session.get("uid"):
+        return redirect("/login")
+    info={}
+    info['wid']=session['wid']
+    info['uid']=session['uid']
+    info['gid']="1"
+    return render_template('waiting.html',data=info)
+
+@app.route("/isCreated", methods=["GET"])
+def isCreated():
+
+    if not session.get("uid"):
+        return redirect("/")
+
+    cur = mysql.connection.cursor()
+    uid = request.args.get('userid')
+
+    empty="SELECT * FROM `group`;"
+    numRow = cur.execute(empty)
+    if(numRow==0):
+        return "false"
+    else:
+        getGrp = "SELECT userid FROM `group` WHERE groupid = ( SELECT groupid FROM `group` AS g WHERE g.userid = %s );"
+        cur.execute(getGrp, [uid])
+        results = cur.fetchall()
+        namelst=[]
+
+        for row in results:
+            getName = "SELECT name FROM `user` WHERE userid= %s"
+            print(row[0])
+            cur.execute(getName,[row[0]])
+            name = cur.fetchall()
+            namelst.append(name[0][0])
+            
+        #print(namelst)
+        participantNames={}
+        id = 0
+        for user in namelst:
+            participantNames[id] = user
+            id = id + 1
+        
+
+        
+        return jsonify(participantNames)
+
+
+
+
+
+@app.route("/clearproto",methods=['POST'])
+def clearproto():
+
+
+    if not session.get("uid"):
+        return redirect("/")
+    cur = mysql.connection.cursor()
+    response=request.get_json()
+    app.logger.info(response)
+    for row in response["data"]:
+        id=response["data"][row].split("myCanvas")[-1]
+        query="delete from protoTable where imgid="+id+";"
+        cur.execute(query)
+    
+    
+    
+    mysql.connection.commit()
+    cur.close()
+    return "Cleared all Proto"
+
+
 
 
 @app.route("/proto")
 def prototype():
-    return render_template('prototype.html')
+
+    if not session.get("uid"):
+        return redirect("/")
+    #*********** Run Sql Query To fetch wid, Gid Uid From DAtabase
+
+    wid = 1
+    gid = 1
+    uid = 2
+
+    data={"workshopid":wid, "groupid": gid, "userid" : uid}
+
+    return render_template('prototype.html',user=data)
 
 
 @app.route("/chat")
 def chatbox():
+
+    if not session.get("uid"):
+        return redirect("/")
+
     return render_template('chatbox.html')
 
 
 @app.route("/define")
 def define():
-    return render_template('define.html')
+
+    if not session.get("uid"):
+        return redirect("/")
+    #*********** Run Sql Query To fetch wid, Gid Uid From DAtabase
+
+    wid = 1
+    gid = 1
+    uid = 2
+
+    data={"workshopid":wid, "groupid": gid, "userid" : uid}
+    return render_template('define.html',user=data)
 
 
 @app.route("/ideate")
 def ideate():
-    return render_template('ideate.html')
+
+    if not session.get("uid"):
+        return redirect("/")
+
+    #*********** Run Sql Query To fetch wid, Gid Uid From DAtabase
+
+    wid = 1
+    gid = 1
+    uid = 2
+
+    data={"workshopid":wid, "groupid": gid, "userid" : uid}
+
+
+    return render_template('ideate.html',user=data)
+
+
+
+@app.route("/finalwall")
+def finalwall():
+
+    if not session.get("uid"):
+        return redirect("/")
+    # global startTime
+    # startTime = int(time.time())
+    wid = 1
+    gid = 1
+    uid = 2
+
+    data={"workshopid":wid, "groupid": gid, "userid" : uid}
+    return render_template('finalwall.html',user=data)
 
 
 
 if __name__ == '__main__':
     app.run(debug=True)
-
